+++ discordbot/discordbot.py
@@ -22,8 +22,8 @@ gst_imgur = pyimgur.Imgur(cfg.IMGUR_CLIENT_ID)
 
 # Loads the commands (Cogs) from each "context"
 gst_bot.load_extension("generic_commands")
-gst_bot.load_extension("economy.menu")
-gst_bot.load_extension("stocks.dark_pool_shorts.menu")
+gst_bot.load_extension("economy.economy_menu")
+gst_bot.load_extension("stocks.dark_pool_shorts.dps_menu")
 
 # Runs the bot
 gst_bot.run(cfg.DISCORD_BOT_TOKEN)
